fix: Suppress TypeScript type error in getByteFrequencyData call
- Replaced explicit type assertion with `@ts-expect-error` directive in `SacredViz.tsx` to address TypeScript's overly strict type inference for `currentAudio.freqs`. This change maintains runtime safety while allowing the build to proceed without errors.

Files Changed:
- apps/integrity-pulse/src/components/SacredViz.tsx

[GI: 0.995] [Consensus: AUREA/ATLAS] [Hash: SHA256]

+++ apps/integrity-pulse/src/components/SacredViz.tsx
@@ -392,7 +392,10 @@ export default function SacredViz({ mockData = true }: SacredVizProps) {
       const t = clock.getElapsedTime();
       const currentAudio = audioContextRef.current;
       if (currentAudio && currentAudio.analyser) {
-        currentAudio.analyser.getByteFrequencyData(currentAudio.freqs as Uint8Array);
+        // TypeScript strict checking: getByteFrequencyData expects Uint8Array<ArrayBuffer>
+        // but infers Uint8Array<ArrayBufferLike>. The array is always ArrayBuffer-backed.
+        // @ts-expect-error - TypeScript incorrectly infers ArrayBufferLike, but runtime is correct
+        currentAudio.analyser.getByteFrequencyData(currentAudio.freqs);
       }
 
       // camera orbit
